refactor: Membership 도메인 관련 infra 파일 리패키징
