+++ backend/src/main/java/com/icehufs/icebreaker/domain/membership/repository/UserRepository.java
@@ -1,9 +1,9 @@
-package com.icehufs.icebreaker.repository;
+package com.icehufs.icebreaker.domain.membership.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.icehufs.icebreaker.entity.User;
+import com.icehufs.icebreaker.domain.membership.domain.entity.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, String> {
